refactor(llmobs): centralize span export in LLMObsProcessor
Move all flush-time LLMObs export into a dedicated LLMObsProcessor wired into
the span aggregator chain, replacing the old apm_filter and sampling fallback
processor. The processor derives whether to drop the APM trace from the export
mode and config (not apm_tracing_enabled or not tracing_enabled) and ships
events via the writer on direct submission.

Restore the _dd.llmobs.submitted tag, set on spans that go through direct
submission via the writer.
